% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{facet_rep_wrap}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lemon}{\code{\link[lemon:facet_rep]{facet_rep_wrap}}}
}}

