% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo.R
\name{theme_glyptodon}
\alias{theme_glyptodon}
\title{Theme Glyptodon}
\usage{
theme_glyptodon(
  base_family = "Roboto Condensed",
  base_size = 11.5,
  plot_title_family = "IBM Plex Sans",
  plot_title_size = base_size + 10.5,
  plot_title_face = "bold",
  plot_title_margin = 5,
  subtitle_family = "Source Sans Pro",
  subtitle_size = base_size + 1,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = "Arial",
  strip_text_size = base_size + 1,
  strip_text_face = "plain",
  strip_background_col = "#835C3B",
  strip_text_col = "#ffffff",
  strip_tran = FALSE,
  caption_family = "Roboto",
  caption_size = base_size - 1,
  caption_face = "italic",
  caption_margin = 10,
  caption_col = "#1c1c1c",
  axis_text_family = "Goldman Sans Condensed",
  axis_text_size = base_size + 2.5,
  axis_title_family = "Arial",
  axis_title_size = base_size,
  axis_col = caption_col,
  axis_title_face = "plain",
  axis_title_just = "rt",
  axis_title_col = caption_col,
  axis = TRUE,
  plot_margin = margin(20, 20, 20, 20),
  plot_background = "#eee8d5",
  legend_position = "top",
  grid_major_col = "#70602d",
  grid_minor_col = "#d7c99d",
  grid = TRUE,
  ticks = TRUE
)
}
\arguments{
\item{ticks}{}
}
\description{
Theme Glyptodon
}
